# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D&D Kids Resources API",
    "description": "API for D&D 5e resource management",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/spells": {
      "get": {
        "tags": [
          "spells"
        ],
        "summary": "List Spells",
        "description": "List all spells with optional filtering.",
        "operationId": "listSpells",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by spell level",
              "title": "Level"
            },
            "description": "Filter by spell level"
          },
          {
            "name": "school",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by spell school",
              "title": "School"
            },
            "description": "Filter by spell school"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Spell"
                  },
                  "title": "Response Listspells"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "spells"
        ],
        "summary": "Create Spell",
        "description": "Create a new spell.",
        "operationId": "createSpell",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpellCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spell"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/spells/{spell_id}/players": {
      "get": {
        "tags": [
          "spells"
        ],
        "summary": "Get Spell Players",
        "description": "Get all players assigned to a spell.",
        "operationId": "getSpellPlayers",
        "parameters": [
          {
            "name": "spell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Spell Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  },
                  "title": "Response Getspellplayers"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "spells"
        ],
        "summary": "Replace Spell Players",
        "description": "Replace all player assignments for a spell.",
        "operationId": "replaceSpellPlayers",
        "parameters": [
          {
            "name": "spell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Spell Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpellPlayerAssignments"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  },
                  "title": "Response Replacespellplayers"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/spells/{spell_id}": {
      "get": {
        "tags": [
          "spells"
        ],
        "summary": "Get Spell",
        "description": "Get a specific spell by ID.",
        "operationId": "getSpell",
        "parameters": [
          {
            "name": "spell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Spell Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spell"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "spells"
        ],
        "summary": "Update Spell",
        "description": "Update an existing spell.",
        "operationId": "updateSpell",
        "parameters": [
          {
            "name": "spell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Spell Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpellUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spell"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "spells"
        ],
        "summary": "Delete Spell",
        "description": "Delete a spell.",
        "operationId": "deleteSpell",
        "parameters": [
          {
            "name": "spell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Spell Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/spells/by-title/{spell_name}": {
      "get": {
        "tags": [
          "spells"
        ],
        "summary": "Get Spell By Title",
        "description": "Get a specific spell by name.",
        "operationId": "getSpellByTitle",
        "parameters": [
          {
            "name": "spell_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spell Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spell"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/players/{player_id}/spells": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Player Spells",
        "description": "Get all spells assigned to a player.",
        "operationId": "getPlayerSpells",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Spell"
                  },
                  "title": "Response Getplayerspells"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "players"
        ],
        "summary": "Replace Player Spells",
        "description": "Replace all spell assignments for a player.",
        "operationId": "replacePlayerSpells",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerSpellAssignments"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Spell"
                  },
                  "title": "Response Replaceplayerspells"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/{player_id}/spells/{spell_id}": {
      "post": {
        "tags": [
          "players"
        ],
        "summary": "Add Spell To Player",
        "description": "Assign a spell to a player.",
        "operationId": "addSpellToPlayer",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          },
          {
            "name": "spell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Spell Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "players"
        ],
        "summary": "Remove Spell From Player",
        "description": "Remove a spell from a player.",
        "operationId": "removeSpellFromPlayer",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          },
          {
            "name": "spell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Spell Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerSpellAssignments": {
        "properties": {
          "spell_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Spell Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PlayerSpellAssignments"
      },
...
      "Spell": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "school": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "School"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "quick_rules": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quick Rules"
          },
          "alternate_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alternate Description"
          },
          "damage": {
            "items": {
              "$ref": "#/components/schemas/SpellDamage"
            },
            "type": "array",
            "title": "Damage"
          },
          "healing": {
            "$ref": "#/components/schemas/SpellHealing"
          },
          "range": {
            "type": "string",
            "title": "Range"
          },
          "higher_levels": {
            "$ref": "#/components/schemas/SpellHigherLevels"
          },
          "casting_times": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Casting Times"
          },
          "duration": {
            "type": "string",
            "title": "Duration"
          },
          "concentration": {
            "type": "boolean",
            "title": "Concentration"
          },
          "ritual": {
            "type": "boolean",
            "title": "Ritual"
          },
          "components": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Components"
          },
          "materials": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materials"
          },
          "attacks": {
            "items": {
              "$ref": "#/components/schemas/SpellAttack"
            },
            "type": "array",
            "title": "Attacks"
          },
          "area_of_effect": {
            "$ref": "#/components/schemas/SpellAreaOfEffect"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "level",
          "description",
          "range",
          "duration",
          "concentration",
          "ritual"
        ],
        "title": "Spell",
        "description": "Response model for the 18-field canonical spell contract."
      },
...
      "SpellError": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "spell_not_found",
              "player_not_found",
              "duplicate_player_ids",
              "failed_to_replace_spell_players",
              "spell_name_already_exists",
              "failed_to_delete_spell"
            ],
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "SpellError"
      },
...
      "SpellPlayerAssignments": {
        "properties": {
          "player_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Player Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SpellPlayerAssignments"
      },
...
    }
  }
}